+++ src/main/java/com/ripple/BE/post/service/ToktokService.java
@@ -10,9 +10,9 @@ import com.ripple.BE.post.dto.CommentListDTO;
 import com.ripple.BE.post.dto.PostDTO;
 import com.ripple.BE.post.dto.PostListDTO;
 import com.ripple.BE.post.exception.PostException;
-import com.ripple.BE.post.repository.PostRepository;
+import com.ripple.BE.post.repository.comment.CommentRepository;
+import com.ripple.BE.post.repository.post.PostRepository;
 import java.time.LocalDate;
-import java.util.Comparator;
 import java.util.List;
 import java.util.Random;
 import lombok.RequiredArgsConstructor;
@@ -20,7 +20,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -31,10 +30,9 @@ import org.springframework.transaction.annotation.Transactional;
 public class ToktokService {
 
     private final PostRepository postRepository;
+    private final CommentRepository commentRepository;
 
     private static final int PAGE_SIZE = 10;
-    private static final String LIKE_COUNT = "likeCount";
-    private static final String CREATED_DATE = "createdDate";
 
     private final PostService postService;
 
@@ -50,6 +48,9 @@ public class ToktokService {
     @Transactional(readOnly = true)
     public PostDTO getToktok(final long id) {
         Post post = postRepository.findById(id).orElseThrow(() -> new PostException(POST_NOT_FOUND));
+        if (post.getUsedDate() == null) {
+            throw new PostException(POST_NOT_FOUND);
+        }
 
         CommentListDTO commentListDTO = getCommentList(post);
 
@@ -59,32 +60,23 @@ public class ToktokService {
     @Transactional(readOnly = true)
     public PostListDTO getToktoks(final int page, final PostSort sort) {
 
-        Sort sortOption =
-                sort == PostSort.POPULAR
-                        ? Sort.by(Sort.Direction.DESC, LIKE_COUNT, CREATED_DATE) // 인기순: 좋아요 수 내림차순, 생성일 내림차순
-                        : Sort.by(Sort.Direction.DESC, CREATED_DATE); // 최신순: 생성일 내림차순
-
-        Pageable pageable = PageRequest.of(page, PAGE_SIZE, sortOption);
+        Pageable pageable = PageRequest.of(page, PAGE_SIZE);
 
         // 게시글 조회 (타입에 따른 필터링)
-        Page<Post> postPage = postRepository.findByTypeAndUsedDateIsNotNull(ECONOMY_TALK, pageable);
+        Page<Post> postPage = postRepository.findUsedToktokPosts(pageable, sort);
 
         return PostListDTO.toPostListDTO(postPage);
     }
 
     private CommentListDTO getCommentList(final Post post) {
-        List<Comment> commentList =
-                post.getCommentList().stream()
-                        .filter(comment -> comment.getParent() == null) // 대댓글이 아닌 경우 필터링
-                        .sorted(Comparator.comparing(Comment::getLikeCount).reversed()) // 좋아요 순으로 정렬
-                        .toList();
+        List<Comment> commentList = commentRepository.findRootCommentsByPost(post);
 
         return CommentListDTO.toCommentListDTO(commentList);
     }
 
     private Post findTodayToktok() {
         return postRepository
-                .findByTypeAndUsedDate(ECONOMY_TALK, LocalDate.now())
+                .findTodayToktokPost(LocalDate.now())
                 .orElseThrow(() -> new PostException(TOKTOK_POST_NOT_FOUND));
     }
 
@@ -96,7 +88,7 @@ public class ToktokService {
         Random random = new Random();
 
         // 오늘 이전에 선정된 게시글을 제외하고 게시글 목록 조회
-        List<Post> unusedPosts = postRepository.findByTypeAndUsedDateIsNull(ECONOMY_TALK);
+        List<Post> unusedPosts = postRepository.findNewToktokPosts();
 
         if (unusedPosts.isEmpty()) {
             throw new PostException(TOKTOK_POST_NOT_FOUND);
